src/main.py
from pydantic import BaseModel
import asyncio
from typing import Dict, List
from fastapi import FastAPI, BackgroundTasks, HTTPException, Request, WebSocket
from fastapi.responses import JSONResponse
from fastapi.staticfiles import StaticFiles
import serial.tools.list_ports
import json
import logging

from src.device_managers import (
    FranklinManager,
    SierraManager, 
    HuaweiManager,
    AndroidManager,
    VoipManager
)
from src.models import Device, SMS
from src.smshub_client import SMSHubClient
from src.database import Database

app = FastAPI()

# ... rest of the code ...

@app.get("/api/ports")
async def list_ports():
    """Get available serial ports"""
    return [port.device for port in serial.tools.list_ports.comports()]

@app.get("/api/devices/{device_id}/stats")
async def get_device_stats(device_id: str):
    """Get detailed device statistics"""
    device = active_devices.get(device_id)
    if not device:
        raise HTTPException(status_code=404, detail="Device not found")
        
    stats = await db.get_device_stats(device_id)
    return stats

@app.get("/api/logs")
async def get_system_logs(level: str = "all", search: str = ""):
    """Get filtered system logs"""
    logs = await db.get_logs(level, search)
    return logs

@app.get("/api/settings")
async def get_settings():
    """Get current system settings"""
    settings = await db.get_settings()
    return settings

@app.post("/api/settings")
async def update_settings(settings: dict):
    """Update system settings"""
    await db.update_settings(settings)
    return {"status": "success"}

# WebSocket for real-time updates
@app.websocket("/ws")
async def websocket_endpoint(websocket: WebSocket):
    await websocket.accept()
    try:
        while True:
            # Send real-time updates
            stats = {
                'active_devices': len(active_devices),
                'messages_today': await db.get_messages_count_today(),
                'system_health': await get_system_health(),
                'queue_size': await db.get_queue_size()
            }
            await websocket.send_json(stats)
            await asyncio.sleep(5)
    except WebSocketDisconnect:
        pass