requirements.txt
fastapi
uvicorn
pyserial
aiohttp
pydantic
python-dotenv
sqlalchemy[asyncio]
asyncpg
alembic
redis>=4.2.0
tenacity
psycopg2-binary